Bikinematics.py split into files in /KivyWidgets
