+++ KivyWidgets/links.kv
@@ -1,0 +1,28 @@
+#:kivy 1.11.1 
+
+<Link>:
+    size_hint:  None,None
+    size: 10,10
+    center: self.midpoint
+    ref: self.a.ref+'_'+self.b.ref
+    canvas.after:
+        Color:
+            rgba: 1,0,0,0.5
+        Rectangle:
+            id: handle
+            size: self.size
+            pos: self.pos
+        Color:
+            rgba: 1,1,1,1
+        Line:
+            points: self.points
+
+<LinkData>:
+    #Coordinate readout box
+    orientation: "horizontal"
+    size_hint_y:None
+    height: sidebar_coords_height
+    Label:
+        text: root.ref
+    TextInput:
+        text: root.len_txt
